refactor two stages
